requirements.txt
requests>=2.28.0
pycryptodome>=3.15.0
pathlib>=1.0.1
rich>=13.0.0
peewee>=3.15.0
PyYAML>=6.0.0
ua-generator>=0.3.0